Firebse Activity moved back to main code
