+++ app/src/androidTest/java/ch/epfl/sdp/contamination/databaseIO/DataReceiverTest.java
@@ -69,6 +69,6 @@ public class DataReceiverTest {
         sleep();
     }
 
-
+    //TODO: test with a non-empty fake grid
 
 }
