+++ app/src/androidTest/java/ch/epfl/sdp/firestore/FirebaseOfflineCacheTest.java
@@ -15,7 +15,6 @@ import org.junit.Rule;
 import org.junit.Test;
 
 import ch.epfl.sdp.TestTools;
-import ch.epfl.sdp.testActivities.FirebaseActivity;
 
 /*
       Run tests in debug mode to see logs & values prints
